requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
amqp==2.6.0               # via kombu
apipkg==1.5               # via execnet
argcomplete==1.11.1       # via zappa
asgiref==3.2.9            # via django
attrs==19.3.0             # via pytest
babel==2.8.0              # via django-phonenumber-field
billiard==3.6.3.0         # via celery
boto3==1.14.10            # via kappa, zappa
botocore==1.17.10         # via boto3, s3transfer
celery==4.4.5             # via -r requirements.in, django-celery-results
certifi==2020.4.5.2       # via requests
cfn-flip==1.2.3           # via troposphere
chardet==3.0.4            # via requests
click==7.1.2              # via cfn-flip, kappa, pip-tools
coverage==5.1             # via -r requirements.in
django-celery-results==1.2.1  # via -r requirements.in
django-cors-headers==3.4.0  # via -r requirements.in
django-money==1.1         # via -r requirements.in
django-phonenumber-field[phonenumbers]==4.0.0  # via -r requirements.in
django==3.0.7             # via -r requirements.in, django-cors-headers, django-money, django-phonenumber-field, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
docutils==0.15.2          # via botocore
durationpy==0.5           # via zappa
execnet==1.7.1            # via pytest-cache
future==0.18.2            # via celery, zappa
hjson==3.0.1              # via zappa
idna==2.9                 # via requests
importlib-metadata==1.6.1  # via argcomplete, kombu, pluggy, pytest
jmespath==0.10.0          # via boto3, botocore, zappa
kappa==0.6.0              # via zappa
kombu==4.6.10             # via celery
mock==4.0.2               # via -r requirements.in
more-itertools==8.4.0     # via pytest
packaging==20.4           # via pytest, pytest-sugar
phonenumbers==8.12.5      # via -r requirements.in, django-phonenumber-field
pip-tools==5.2.1          # via zappa
placebo==0.9.0            # via kappa
pluggy==0.13.1            # via pytest
postcode-validator-uk==0.0.5  # via -r requirements.in
py-moneyed==0.8.0         # via django-money
py==1.8.2                 # via pytest
pycodestyle==2.6.0        # via -r requirements.in
pyparsing==2.4.7          # via packaging
pytest-cache==1.0         # via -r requirements.in
pytest-django==3.9.0      # via -r requirements.in
pytest-sugar==0.9.3       # via -r requirements.in
pytest==5.4.3             # via pytest-cache, pytest-django, pytest-sugar
python-dateutil==2.6.1    # via botocore, zappa
python-slugify==4.0.0     # via zappa
pytz==2020.1              # via babel, celery, django
pyyaml==5.3.1             # via cfn-flip, kappa, zappa
requests==2.24.0          # via -r requirements.in, zappa
s3transfer==0.3.3         # via boto3
six==1.15.0               # via cfn-flip, packaging, pip-tools, python-dateutil, zappa
sqlparse==0.3.1           # via django
termcolor==1.1.0          # via pytest-sugar
text-unidecode==1.3       # via python-slugify
toml==0.10.1              # via zappa
tqdm==4.46.1              # via zappa
troposphere==2.6.1        # via zappa
urllib3==1.25.9           # via botocore, requests
vine==1.3.0               # via amqp, celery
wcwidth==0.2.4            # via pytest
werkzeug==0.16.1          # via zappa
wheel==0.34.2             # via zappa
wsgi-request-logger==0.4.6  # via zappa
zappa==0.51.0             # via -r requirements.in
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
